Add color-v2 package and remove \color from base package.  The latex color package is autoloaded unless the page author include color-v2 in the package list instead (or adjusts the autoload definitions).
