+++ components/src/input/tex/extensions/color-v2/build.json
@@ -1,0 +1,5 @@
+{
+    "component": "input/tex/extensions/color-v2",
+    "targets": ["input/tex/color_v2"]
+}
+
