+++ components/src/input/tex/extensions/color-v2/color-v2.js
@@ -1,0 +1,1 @@
+import './lib/color-v2.js';
