+++ components/src/input/tex/extensions/color-v2/webpack.config.js
@@ -1,0 +1,11 @@
+const PACKAGE = require('../../../../../webpack.common.js');
+
+module.exports = PACKAGE(
+    'input/tex/extensions/color-v2',    // the package to build
+    '../../../../../../mathjax3',       // location of the mathjax3 library
+    [                                   // packages to link to
+        'components/src/input/tex-base/lib',
+        'components/src/core/lib'
+    ],
+    __dirname                           // our directory
+);
